Renamed maths.combinatorics.core to .numbers for clarity, absorbing .graphenum in the process
